fix(newsletters): address post-commit review findings
Extracts the send + analytics business logic from the controller into a
new NewsletterService, and hardens several edge cases surfaced by the
post-commit review pair.

Code-reviewer findings addressed:
- Critical: business logic in controller. The five private methods
  (dispatchNewsletter, fanOutEmails, buildEmptyAnalytics,
  buildAnalyticsFromRecords, aggregateDailyOpens) move to a new
  services/newsletter.service.ts so the three-file pattern is honored.
  The controller is back to validation + logging lifecycle + delegation.
- Critical (partial): duplicate-send risk on markSent failure. Adds a
  3-attempt exponential-backoff retry around the markSent call (only
  retries 5xx; 4xx errors surface immediately). This rides out the
  transient network / upstream 5xx case. Full operator-retry idempotency
  still requires the Go service to support persisting a group_id on the
  draft BEFORE the fan-out — tracked as a follow-up; the comment in
  dispatchNewsletter documents the gap.
- Important: getStatusByGroup 502 on empty/missing records would surface
  as a failed analytics page. getAnalytics now catches the
  EMAIL_SERVICE_ERROR MicroserviceError and falls through to the empty
  analytics payload, with a warning log carrying the underlying message.
- Important: totalOpens/uniqueOpens vs dailyOpens used different
  predicates. Both now use `r.opened && r.opened_at` so the headline
  count always sums to the chart buckets.
- Important: the email-service `{error}` envelope check was too broad
  (`'error' in parsed`) — a successful record with a legitimate `error`
  field would be misclassified. Tightened to require a single-key object
  with a string `error` field.
- Important: logger.error inside dispatchNewsletter ran from the
  controller catch chain. Now using logger.warning per attempt in the
  service; the central apiErrorHandler logs the final error.
- Important: `payload as CreateNewsletterDraftRequest` cast in send().
  Construct the create-draft body explicitly from validated fields so a
  future divergence between NewsletterSendPayload and
  CreateNewsletterDraftRequest surfaces as a type error.

Learnings-reviewer findings addressed:
- Important: recipient email leaking into structured logs. Dropped `to`
  from the email-service debug-send log and from the test-send success
  log. group_id remains as the correlation key (no PII).

Files:
- services/newsletter.service.ts: new — owns dispatchNewsletter,
  getAnalytics, plus markSentWithRetry / isRetryable / sleep,
  fanOutEmails, buildEmptyAnalytics, buildAnalyticsFromRecords,
  aggregateDailyOpens.
- services/email-service.client.ts: tightened envelope detection
  (isErrorEnvelope helper); dropped recipient email from debug log.
- controllers/newsletter.controller.ts: delegates send / sendDraft /
  getAnalytics to NewsletterService. testSend success log no longer
  records the recipient email. send() constructs
  CreateNewsletterDraftRequest explicitly.
