+++ apps/lfx-one/src/server/controllers/newsletter.controller.ts
@@ -4,33 +4,24 @@
 import { NEWSLETTER_RAW_CONTENT_MAX_LENGTH, NEWSLETTER_SYSTEM_PROMPT_MAX_LENGTH } from '@lfx-one/shared/constants';
 import {
   CreateNewsletterDraftRequest,
-  EmailRecipientRecord,
   GenerateNewsletterRequest,
-  Newsletter,
-  NewsletterAnalytics,
   NewsletterContextType,
-  NewsletterDailyOpens,
   NewsletterListParams,
-  NewsletterRecipient,
   NewsletterRecipientCountPayload,
-  NewsletterSendFailure,
   NewsletterSendPayload,
-  NewsletterSendResult,
   NewsletterStatus,
   NewsletterTestSendPayload,
   UpdateNewsletterDraftRequest,
 } from '@lfx-one/shared/interfaces';
 import { stripHtml } from '@lfx-one/shared/utils';
-import { randomUUID } from 'crypto';
 import { NextFunction, Request, Response } from 'express';
 
 import { ServiceValidationError } from '../errors';
 import { AiService } from '../services/ai.service';
 import { EmailServiceClient } from '../services/email-service.client';
 import { logger } from '../services/logger.service';
+import { NewsletterService } from '../services/newsletter.service';
 import { NewsletterServiceClient } from '../services/newsletter-service.client';
-
-const EMAIL_SEND_CONCURRENCY = 5;
 
 const VALID_CONTEXT_TYPES = new Set(['foundation', 'project']);
 const SUBJECT_MAX_LENGTH = 200;
@@ -41,6 +32,7 @@ const CONTEXT_NAME_MAX_LENGTH = 200;
 export class NewsletterController {
   private newsletterClient: NewsletterServiceClient = new NewsletterServiceClient();
   private emailServiceClient: EmailServiceClient = new EmailServiceClient();
+  private newsletterService: NewsletterService = new NewsletterService(this.newsletterClient, this.emailServiceClient);
   private aiService: AiService = new AiService();
 
   /**
@@ -121,7 +113,8 @@ export class NewsletterController {
         text: stripHtml(payload.bodyHtml),
       });
 
-      logger.success(req, 'newsletter_test_send', startTime, { to: payload.toEmail });
+      // PII (recipient email) intentionally omitted from log metadata.
+      logger.success(req, 'newsletter_test_send', startTime, {});
       res.json({ ok: true });
     } catch (error) {
       next(error);
@@ -148,9 +141,19 @@ export class NewsletterController {
       this.validateCommonPayload(payload, req.path, 'newsletter_send');
       this.validateCommitteeUids(payload.committeeUids, req.path, 'newsletter_send');
 
-      const draft = await this.newsletterClient.createDraft(req, payload as CreateNewsletterDraftRequest);
+      // Build the create-draft request explicitly so we keep type safety even
+      // if NewsletterSendPayload and CreateNewsletterDraftRequest diverge.
+      const createDraftRequest: CreateNewsletterDraftRequest = {
+        contextType: payload.contextType,
+        contextUid: payload.contextUid,
+        subject: payload.subject,
+        bodyHtml: payload.bodyHtml,
+        edReplyEmail: payload.edReplyEmail,
+        committeeUids: payload.committeeUids,
+      };
+      const draft = await this.newsletterClient.createDraft(req, createDraftRequest);
 
-      const result = await this.dispatchNewsletter(req, draft, 'newsletter_send');
+      const result = await this.newsletterService.dispatchNewsletter(req, draft, 'newsletter_send');
 
       logger.success(req, 'newsletter_send', startTime, {
         newsletter_id: draft.id,
@@ -301,7 +304,7 @@ export class NewsletterController {
       const version = parseIfMatch(req);
 
       const draft = await this.newsletterClient.getDraft(req, id);
-      const result = await this.dispatchNewsletter(req, draft, 'newsletter_send_draft', version);
+      const result = await this.newsletterService.dispatchNewsletter(req, draft, 'newsletter_send_draft', version);
 
       logger.success(req, 'newsletter_send_draft', startTime, {
         draft_id: id,
@@ -392,21 +395,7 @@ export class NewsletterController {
       }
 
       const newsletter = await this.newsletterClient.getDraft(req, id);
-
-      let analytics: NewsletterAnalytics;
-      if (!newsletter.groupId) {
-        // Fallback for newsletters sent before this integration shipped, and
-        // for drafts (which legitimately have no engagement data). Returns an
-        // empty NewsletterAnalytics payload so the analytics page renders.
-        logger.debug(req, 'newsletter_analytics', 'Newsletter has no groupId — returning empty analytics', {
-          newsletter_id: id,
-          status: newsletter.status,
-        });
-        analytics = this.buildEmptyAnalytics(newsletter);
-      } else {
-        const records = await this.emailServiceClient.getStatusByGroup(req, newsletter.groupId);
-        analytics = this.buildAnalyticsFromRecords(newsletter, records);
-      }
+      const analytics = await this.newsletterService.getAnalytics(req, newsletter);
 
       logger.success(req, 'newsletter_analytics', startTime, {
         newsletter_id: id,
@@ -482,205 +471,6 @@ export class NewsletterController {
     } catch (error) {
       next(error);
     }
-  }
-
-  /**
-   * Per-recipient email-service fan-out + sent-state persistence.
-   *
-   * Resolves recipients from the newsletter's committeeUids, mints a UUID
-   * group_id, dispatches one send_email per recipient with bounded
-   * concurrency, then PATCHes the newsletter to status=sent unless every
-   * recipient failed (matches the previous behavior of not flipping status
-   * on total failure).
-   *
-   * Validations BEFORE any email goes out (cheaper to fail here than to
-   * dispatch and then discover the Go service will reject the markSent):
-   *   - Already-sent newsletters are rejected to prevent duplicate-send
-   *     races (e.g., double-clicking "Send now" while the first request
-   *     is in flight).
-   *   - Empty recipient lists are rejected — the UI shouldn't allow this,
-   *     but a committee with zero resolvable members would otherwise mint
-   *     a group_id, send no emails, and silently no-op.
-   */
-  private async dispatchNewsletter(req: Request, newsletter: Newsletter, operation: string, ifMatchVersion?: number): Promise<NewsletterSendResult> {
-    if (newsletter.status === 'sent') {
-      throw ServiceValidationError.forField('status', 'Newsletter has already been sent', {
-        operation,
-        service: 'newsletter_controller',
-        path: req.path,
-      });
-    }
-
-    const recipientsResponse = await this.newsletterClient.getRecipients(req, { committeeUids: newsletter.committeeUids });
-    const recipients = recipientsResponse.recipients;
-
-    if (recipients.length === 0) {
-      throw ServiceValidationError.forField('committeeUids', 'Selected committees resolved to zero recipients; nothing to send', {
-        operation,
-        service: 'newsletter_controller',
-        path: req.path,
-      });
-    }
-
-    const groupId = randomUUID();
-
-    logger.debug(req, operation, 'Dispatching newsletter via email-service', {
-      newsletter_id: newsletter.id,
-      group_id: groupId,
-      recipient_count: recipients.length,
-    });
-
-    const { sent, failures } = await this.fanOutEmails(req, newsletter, recipients, groupId);
-
-    if (sent > 0) {
-      const versionToUse = ifMatchVersion ?? newsletter.version;
-      try {
-        await this.newsletterClient.markSent(req, newsletter.id, { groupId, ifMatchVersion: versionToUse });
-      } catch (error) {
-        // Emails went out but the status flip failed. Log loudly so the
-        // group_id stays recoverable from logs and the operator can retry.
-        logger.error(req, operation, Date.now(), error, {
-          newsletter_id: newsletter.id,
-          group_id: groupId,
-          sent,
-          failed: failures.length,
-          message: 'Emails delivered but newsletter mark-sent PATCH failed; group_id captured in logs for manual recovery',
-        });
-        throw error;
-      }
-    }
-
-    return {
-      totalRecipients: recipients.length,
-      sent,
-      failed: failures.length,
-      failures,
-      groupId,
-    };
-  }
-
-  /**
-   * Sends one email per recipient with a bounded number of in-flight
-   * requests. NATS round-trips are fast (~25ms in dev) but a fully parallel
-   * Promise.all would still saturate the connection for large committees.
-   */
-  private async fanOutEmails(
-    req: Request,
-    newsletter: Pick<Newsletter, 'subject' | 'bodyHtml'>,
-    recipients: NewsletterRecipient[],
-    groupId: string
-  ): Promise<{ sent: number; failures: NewsletterSendFailure[] }> {
-    const text = stripHtml(newsletter.bodyHtml);
-    const failures: NewsletterSendFailure[] = [];
-    let sent = 0;
-    let cursor = 0;
-
-    const worker = async (): Promise<void> => {
-      while (cursor < recipients.length) {
-        const index = cursor++;
-        const recipient = recipients[index];
-
-        try {
-          await this.emailServiceClient.sendEmail(req, {
-            to: recipient.email,
-            subject: newsletter.subject,
-            html: newsletter.bodyHtml,
-            text,
-            group_id: groupId,
-          });
-          sent++;
-        } catch (error) {
-          failures.push({
-            email: recipient.email,
-            reason: error instanceof Error ? error.message : 'Unknown error',
-          });
-        }
-      }
-    };
-
-    const workerCount = Math.min(EMAIL_SEND_CONCURRENCY, recipients.length);
-    await Promise.all(Array.from({ length: workerCount }, () => worker()));
-
-    return { sent, failures };
-  }
-
-  /**
-   * Empty-shape analytics for newsletters that don't have a groupId yet —
-   * drafts, or sent newsletters that predate this integration.
-   */
-  private buildEmptyAnalytics(newsletter: Newsletter): NewsletterAnalytics {
-    return {
-      newsletterId: newsletter.id,
-      subject: newsletter.subject,
-      status: newsletter.status,
-      sentAt: newsletter.sentAt,
-      totalRecipients: 0,
-      delivered: 0,
-      failed: 0,
-      totalOpens: 0,
-      uniqueOpens: 0,
-      openRate: 0,
-      dailyOpens: [],
-    };
-  }
-
-  /**
-   * Aggregate per-recipient EmailRecipientRecords into the existing
-   * NewsletterAnalytics shape so the frontend (which expects daily-opens
-   * chart data) keeps working unchanged.
-   *
-   * Note: the email-service tracks "opened" as a boolean per recipient (no
-   * multi-open count), so totalOpens == uniqueOpens for our purposes.
-   */
-  private buildAnalyticsFromRecords(newsletter: Newsletter, records: EmailRecipientRecord[]): NewsletterAnalytics {
-    const totalRecipients = records.length;
-    const delivered = records.filter((r) => r.delivered).length;
-    const failed = records.filter((r) => r.failed).length;
-    const opensCount = records.filter((r) => r.opened).length;
-    const openRate = delivered > 0 ? opensCount / delivered : 0;
-
-    const dailyOpens = this.aggregateDailyOpens(records);
-
-    const eventTimestamps: number[] = [];
-    for (const record of records) {
-      if (record.opened_at) {
-        eventTimestamps.push(new Date(record.opened_at).getTime());
-      }
-      if (record.delivered_at) {
-        eventTimestamps.push(new Date(record.delivered_at).getTime());
-      }
-    }
-    const lastEventAt = eventTimestamps.length > 0 ? new Date(Math.max(...eventTimestamps)).toISOString() : undefined;
-
-    return {
-      newsletterId: newsletter.id,
-      subject: newsletter.subject,
-      status: newsletter.status,
-      sentAt: newsletter.sentAt,
-      totalRecipients,
-      delivered,
-      failed,
-      totalOpens: opensCount,
-      uniqueOpens: opensCount,
-      openRate,
-      dailyOpens,
-      lastEventAt,
-    };
-  }
-
-  private aggregateDailyOpens(records: EmailRecipientRecord[]): NewsletterDailyOpens[] {
-    const buckets = new Map<string, number>();
-    for (const record of records) {
-      if (!record.opened || !record.opened_at) {
-        continue;
-      }
-      const day = record.opened_at.slice(0, 10);
-      buckets.set(day, (buckets.get(day) ?? 0) + 1);
-    }
-
-    return Array.from(buckets.entries())
-      .map(([date, opens]) => ({ date, opens, uniqueOpens: opens }))
-      .sort((a, b) => a.date.localeCompare(b.date));
   }
 
   private validateCommonPayload(
